requirements.txt
numpy==1.17.2
torch==1.0.1
spacy==2.2.3
termcolor==1.1.0
colorama==0.4.1
matplotlib==3.1.1
pandas==1.0.3
qwikidata==0.4.0
nltk==3.5
astropy==4.0.1.post1
tabulate==0.8.7
aiohttp==3.6.2
newspaper3k==0.2.8
transformers
